# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PublicLibrary API",
    "version": "1.0.0",
    "description": "API for PublicLibrary domain"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/book/registerbook": {
      "post": {
        "summary": "Execute RegisterBook command",
        "operationId": "RegisterBook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterBookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/book/{bookId}/markbookasrestricted": {
      "post": {
        "summary": "Execute MarkBookAsRestricted command",
        "operationId": "MarkBookAsRestricted",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkBookAsRestrictedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkBookAsRestrictedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/book/{bookId}": {
      "get": {
        "summary": "Get Book by ID",
        "operationId": "getBookById",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
...
    "/fee/accruelatefee": {
      "post": {
        "summary": "Execute AccrueLateFee command",
        "operationId": "AccrueLateFee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccrueLateFeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccrueLateFeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/fee/{feeId}/makepartialpayment": {
      "post": {
        "summary": "Execute MakePartialPayment command",
        "operationId": "MakePartialPayment",
        "parameters": [
          {
            "name": "feeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakePartialPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakePartialPaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/fee/{feeId}/payremainingbalance": {
      "post": {
        "summary": "Execute PayRemainingBalance command",
        "operationId": "PayRemainingBalance",
        "parameters": [
          {
            "name": "feeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRemainingBalanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayRemainingBalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/fee/{feeId}": {
      "get": {
        "summary": "Get Fee by ID",
        "operationId": "getFeeById",
        "parameters": [
          {
            "name": "feeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fee found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fee"
                }
              }
            }
          },
          "404": {
            "description": "Fee not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "restricted": {
            "type": "boolean"
          }
        },
        "required": [
          "bookId",
          "title",
          "restricted"
        ]
      },
...
      "Fee": {
        "type": "object",
        "properties": {
          "feeId": {
            "type": "string",
            "format": "uuid"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          },
          "loanId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "accruedDate": {
            "type": "string"
          },
          "paidAmount": {
            "type": "number",
            "format": "decimal"
          }
        },
        "required": [
          "feeId",
          "patronId",
          "loanId",
          "amount",
          "accruedDate",
          "paidAmount"
        ]
      },
      "RegisterBookRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "restricted": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "restricted"
        ]
      },
      "RegisterBookResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "MarkBookAsRestrictedRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "bookId",
          "bookId"
        ]
      },
      "MarkBookAsRestrictedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "AccrueLateFeeRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "decimal"
          }
        },
        "required": [
          "loanId",
          "amount"
        ]
      },
      "AccrueLateFeeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "feeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "MakePartialPaymentRequest": {
        "type": "object",
        "properties": {
          "feeId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "decimal"
          }
        },
        "required": [
          "feeId",
          "feeId",
          "amount"
        ]
      },
      "MakePartialPaymentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "feeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "PayRemainingBalanceRequest": {
        "type": "object",
        "properties": {
          "feeId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "decimal"
          }
        },
        "required": [
          "feeId",
          "feeId",
          "amount"
        ]
      },
      "PayRemainingBalanceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "feeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      }
    }
  }
}